// logicommerce/types.ts
export type LogicommerceProductSorts =
  | "id.asc"
  | "id.desc"
  | "pId.asc"
  | "pId.desc"
  | "sku.asc"
  | "sku.desc"
  | "name.asc"
  | "name.desc"
  | "priority.asc"
  | "priority.desc"
  | "price.asc"
  | "price.desc"
  | "offer.asc"
  | "offer.desc"
  | "featured.asc"
  | "featured.desc"
  | "publicationDate.asc"
  | "publicationDate.desc";